+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.9.2](https://github.com/realChakrawarti/ingest/compare/v1.9.1...v1.9.2) (2025-08-31)
+
+
+### Bug Fixes
+
+* subreddit post height, download when cobalt instances are available & play w/ default audio ([45f420b](https://github.com/realChakrawarti/ingest/commit/45f420b4377ba5caf965f28ec87b84d2b70659c8))
+* update cobalt instance api endpoint ([3643630](https://github.com/realChakrawarti/ingest/commit/3643630df0fa6097027efa68f309b515c95f85f1))
+
 ### [1.9.1](https://github.com/realChakrawarti/ingest/compare/v1.9.0...v1.9.1) (2025-08-14)
 
 ## [1.9.0](https://github.com/realChakrawarti/ingest/compare/v1.8.19...v1.9.0) (2025-08-14)